latex: code highlighting debugged
Highlighted Sourcecode from Pygments requires the fancyvrb and color packages.
To get it working with the Sphinx LaTeX Parser some commands where necessary
which were commented out. And last but not least creating a lstlisting
environment is not a reasonable choice, considering that this package has the
same job as Pygments. Therefore the overwritten method depart_literal_block was
removed, that Sphinx' default will be used. The result is for some reasons
rather ugly, but at least it works now.

I also changed the code example to a Python function to show that highlighting
really works now.

FYI: For some reason I also needed to uncomment the `notice` environment
defintion. Building from master without the changes introduced here didn't work
at all. Might be something I did, though.
But maybe this is a hint that a lot more of the sphinx.sty needs to be
uncommented, but nobody noticed because it wasn't used until now.

Fixes #8

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>